[FEAT] effectmachine추가, actorstatobject를 actor 클래스에 추가, base_stat을 int형에서 float형으로 변환, playerstatobject에 effectmachine 추가

+++ Assets/Scripts/Actor/ActorStatObject.cs
@@ -20,10 +20,10 @@ namespace Actor
     [CreateAssetMenu(fileName = "Actor Data", menuName = "Scriptable Object/Actor Data")]
     public class ActorStatObject : ScriptableObject
     {
-        public int health;
-        public int atkPower;
-        public int speed;
-        public int defence;
+        public float health;
+        public float atkPower;
+        public float speed;
+        public float defence;
         public ElementalType element;
     }
 }
